+++ src/main/java/net/silentchaos512/powerscale/PowerScale.java
@@ -1,6 +1,5 @@
 package net.silentchaos512.powerscale;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
@@ -18,7 +17,8 @@ import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.AddReloadListenerEvent;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
-import net.silentchaos512.powerscale.setup.*;
+import net.silentchaos512.powerscale.setup.PsRegistries;
+import net.silentchaos512.powerscale.setup.Registration;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;