+++ src/main/java/net/silentchaos512/powerscale/data/loot/EntityLootTables.java
@@ -13,6 +13,7 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceWit
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
 import net.silentchaos512.powerscale.core.EntityGroups;
 import net.silentchaos512.powerscale.loot.condition.CrystalDropChanceConfig;
+import net.silentchaos512.powerscale.loot.condition.CrystalDropWithCooldown;
 import net.silentchaos512.powerscale.loot.condition.MobProperties;
 import net.silentchaos512.powerscale.setup.PsItems;
 
@@ -35,6 +36,7 @@ public class EntityLootTables implements LootTableSubProvider {
                                 // Always with random chance
                                 LootPool.lootPool()
                                         .when(CrystalDropChanceConfig.configWithLootingBonus(registries))
+                                        .when(CrystalDropWithCooldown.builder())
                                         .when(LootItemKilledByPlayerCondition.killedByPlayer())
                                         .add(
                                                 LootItem.lootTableItem(PsItems.HEART_CRYSTAL)